Moved reload and save functionality tests to own files
